test: consolidate duplicated spawn fixture and orphan test files (closes #454)
The binary and db test layers carried redundant single-test files and a
duplicated process-spawn fixture.

musefs-db: fold the two orphan single-test binaries into the existing
suites that already exercise the same subsystem — the `Format` enum
round-trip into `tracks.rs` and the connection `data_version()` watermark
into `triggers.rs`. Delete `format.rs` and `change_detection.rs`.

musefs: merge `env_config.rs` into `cli_process.rs` so the `env_clear()`
spawn helper, FLAC fixture, and `unopenable_db` helper exist once. The
CLI-flag and `MUSEFS_*` env triggers share one process-boundary contract.
Drop `valid_boolean_env_is_accepted` (`"true"`), fully subsumed by
`boolish_boolean_env_values_are_accepted` (whose table includes `"TRUE"`).
