+++ musefs-db/tests/tracks.rs
@@ -230,3 +230,15 @@ fn delete_changelog_through_for_test_prunes_the_prefix() {
         "rows through max_seq - 1 must actually be deleted"
     );
 }
+
+#[test]
+fn format_round_trips_through_db_string() {
+    assert_eq!(Format::Flac.as_str(), "flac");
+    assert_eq!(Format::Mp3.as_str(), "mp3");
+    assert_eq!("flac".parse::<Format>(), Ok(Format::Flac));
+    assert_eq!("mp3".parse::<Format>(), Ok(Format::Mp3));
+    assert_eq!(
+        "ogg".parse::<Format>(),
+        Err(strum::ParseError::VariantNotFound)
+    );
+}
